Use postgres database for local development
Add postgres container to docker-compose.yml. This loads the
password for the default postgres user from the environment. Update
the README to indicate this should be defined within a .env file
in the root directory.

Replace the H2 dependency with one for postgres and update the
dev spring profile to use the postgres dialect. Load the password
from the .env file used to configure docker compose.