Send messages on checkin invite and submission
Message templates have been added to the gov.uk notify configuration
for checkin invite and submission messages. Add these to the
app configuration(s) and remove the templates used for the old invite
flow.

Add a Message interface type for messages which can be sent to a
recipient. A recipient is a type which exposes a collection of
contact methods (phone number or email). Implement the Contactable
inteface for Offender and Practitioner types.

Add message types for the checkin invite and submission messages.
These return a collection of named template parameters based on the
message properties and application configuration. The application
parameters are exposed via a new AppConfig class which is used to
resolve references to checkins within the UI.

Update the NotificationService to send a generic message to an
arbitrary contactable recipient. The template is resolved from
the configuration based on the message and notification method
types.