Merge local profile into dev profile
The dev and local profiles are both used when developing locally
so merge them into a single profile. Where one hard-codes a
config setting and the other uses the environment, the envrionment
method has been retained.

Merge the local profile config into the dev config and remove the
application-local.yml file. Rename LocalAwsConfig to DevAwsConfig
and update the 'rekog' profile to configure the AWS endpoint when
running with the dev profile.