Merge local and dev profiles (#30)
* Create typed configuration for notification message templates

Add MessageTemplateConfig type which maps the notify message template
configuration. This contains the template configuration for each
contact method within the application.

Remove the explicit construction of configuration beans from the
existing notify config class.

* Merge local profile into dev profile

The dev and local profiles are both used when developing locally
so merge them into a single profile. Where one hard-codes a
config setting and the other uses the environment, the envrionment
method has been retained.

Merge the local profile config into the dev config and remove the
application-local.yml file. Rename LocalAwsConfig to DevAwsConfig
and update the 'rekog' profile to configure the AWS endpoint when
running with the dev profile.

---------

Co-authored-by: Lee Kitching <lee.kitching@justice.gov.uk>