Add next checkin date and checkin frequency to Offender model
Add date of next checkin to the Offender model along with a duration
for the checkin frequency. At present only weekly, bi-weekly and
every four weeks are supported within the UI and API. Add a
CheckinFrequency enum for the supported cases and map to/from this
type for persistance. The checkin duration is mapped to an interval
column type within postgres.