flag survey responses (#36)
The flagged responses are exposed on the checkin DTO instances
as a computed property: `flaggedResponses`.

From now on we will expect survey responses to contain a "version"
entry. Function that implements what to flag/not flag will need to
be optionally added to a mapping `versionToFlaggingFn`. If no such
function is found (or no version string is found in the survey)
`flaggedResponses` will be an empty list.