Scheduling notifications
The class uses the introduced scheduler lock to prevent from more
than one instance of the job running.

The job goes over all relevant offender records, checks if it's time
for a checkin, then creates the checkin (and that tirggers
a notification).

Meant to be uses by practitioners/admins to send notifications manually.

Makes it less awkward to handle the dates in code

The notification IDs are then stored in
`OffenderCheckin.notifications` to enable us to follow up on the
notification status (that part is not implemented yet, it would probably
be another scheduled job).