Add next checkin date and checkin frequency to Offender model (#37)
* Add next checkin date and checkin frequency to Offender model

Add date of next checkin to the Offender model along with a duration
for the checkin frequency. At present only weekly, bi-weekly and
every four weeks are supported within the UI and API. Add a
CheckinFrequency enum for the supported cases and map to/from this
type for persistance. The checkin duration is mapped to an interval
column type within postgres.

* Scheduling notifications

The class uses the introduced scheduler lock to prevent from more
than one instance of the job running.

The job goes over all relevant offender records, checks if it's time
for a checkin, then creates the checkin (and that tirggers
a notification).

Meant to be uses by practitioners/admins to send notifications manually.

Makes it less awkward to handle the dates in code

The notification IDs are then stored in
`OffenderCheckin.notifications` to enable us to follow up on the
notification status (that part is not implemented yet, it would probably
be another scheduled job).

Other changes:

* We can now create checkins with due date of today
* set notifier schedule to run every 15 minutes
* move schedule spec to config
* exception handler for invalid requests
* more explicit notification references
* sending notification now requires to pass a context
* clock uses Europe/London TZ
* fix for checkin submission - we didn't allow for the 72h window
to submit checkins.
* pull config value directly to service object
* time zone info is obtained from the Clock bean
* update http script to handle upload locations correctly
-------

Co-authored-by: Lee Kitching <lee.kitching@justice.gov.uk>
Co-authored-by: Roland Sadowski <szabla@gmail.com>