Define local practitioner in configuration
Create a new local profile containing local practitioner users.
When this profile is active an in-memory repository of the
listed practitioners is used instead of the configured 'manage
users' API client.

Move the test practitioners to a test configuration for the
practitioners repository and fix the tests. Note the verifications
on the mock notification service have been temporarily disabled.