Fix checkin day calculation in invites job
Fixes ESUP-607 - Prevent checkin invites being created before a
PoP's first checkin date.

Fix calculation of checkin days based on the PoP first checkin
date and interval. Period.toDays() returns only the residual
number of days in the period, not the total number. So for example
a 6-week period (42 days) is represented as 1 month and 11 days.
The toDays() method returns 11 in this case, not 42 as required.
Use LocalDate.until to get the total number of days between the
first checkin date and job execution date.

Add tests for calculating checkin days to CheckinNotifierTest.