Use Postgres in a test container instead of H2
Issue ESUP-636 - Add testcontainers dependency and use it to run
a Postgres container for the duration of the tests. Remove the
test H2 configuration so the Postgres configuration from the default
Spring profile is used. The URL, database and password for the
data source are provided by the IntegrationTestBase which
dynamically registers the properties from the running container.

The container is started before the first test is executed and
stopped after the last test run. This is done via a listener
for the JUnit test session which is registered as a service
provider.