update notif. status updater job to include one-off notif. statuses (#76)
Modifies the notification status update job to also fetch the status for one off notifications. One off notifications are saved with a reference of 'CHK-YYYY-MM-dd' and we use that to fetch one-offs sent on a particular day.

Also: save checkin notification in createCheckin when called with a 'single' notification context.

---------

Co-authored-by: Roland Sadowski <roland.sadowski@justice.gov.uk>