Move stub id verification config into its own class
Add StubRekogConfig class which defines the beans required for
stub id verification. Snapshot images are sent to a localstack s3
bucket and the verification process always succeeds. This is only
active during the tests or when the stubrekog profle is specified.

Add configuration for the stub verification process to the README.

Remove the old TestRekogConfig which is no longer needed since
the stub config is used there instead.