add 'stubrekog' profile/config with a stub face comparison service (#80)
* add 'stubrekog' profile/config with a stub implementation

This allows to run the application without relying on real rekognition
(and associated auth configuration).

* Move stub id verification config into its own class

Add StubRekogConfig class which defines the beans required for
stub id verification. Snapshot images are sent to a localstack s3
bucket and the verification process always succeeds. This is only
active during the tests or when the stubrekog profle is specified.

Remove the old TestRekogConfig which is no longer needed since
the stub config is used there instead.

---------

Co-authored-by: Roland Sadowski <roland.sadowski@justice.gov.uk>
Co-authored-by: Lee Kitching <lee.kitching@justice.gov.uk>