WIP: per site stats
Adds a repository class which populates a temp table for subsequent
analytics queries. This is a temporary solution - we need the
practitioners+site data to be in the database. Right now it's just
a few dozens of practitioners, but possibly thousands in the future
(there's ~20k practitioners in MOJ at the moment, we would need to
handle a subset of that).

The queries live in 'resources/db/queries' and results are returned
in one big wrapper object. We should think of caching the results
(we need to agree on how fresh the data needs to be, a few hours
behind shouldn't bother anyone?).