Tracking delivery of non-invite notifications (#101)
Adds a GenericNotification entity for tracking status of notifications
other than offender checkins along with a job that updates the status
periodically.

The stats (implementation will come in another PR) that this PR enables
us to gather won't be complete (GOV UK Notify holds that information 
only for seven days) and I decided not to complicate the implementation 
with backfilling the records for the past week. 

Also:
- updated the Referencables to return `reference` strings more
  specific to the message they're associated with (and not just
  XYZ-$UUID), and to include a date.
- moved a few Notification* classes to 'notification' package

Note: I did deploy the branch in dev and verified the job works.