Di/esup 729 avg review response time metric (#105)
Calculating the review response time as the time between the checkin being submitted and the review being completed

Returning the time interval & time interval as a string in the format 0h0m0s per site and a total time interval string to avoid calculations in the FE

There is a corresponding FE change that is dependant on these changes

--- UPDATES ---

The PR review highlighted two issues with this PR

The total average was an average of averages rather than the true total average of all reviewed check ins. This has been fixed in the latest change added to this PR.

The submittedAt & reviewedAt fields in the offender_checkin table aren't indexed. Having had a look, this is true of a number of fields used for stats queries. To avoid lots of small migrations I will create a new ticket to add any new indexes together.