nginx rate limits for test, preprod, prod
The dev/test numbers have been tested with some load testing.
Admittedly, the dev/test setup is not representative, as the data
is not close to what exists in prod (we don't have a data generation
code yet, and I don't want to get sidetracked). Also, hardware specs
are different in dev/test than in prod/preprod.