Add V1 endpoints (#121)
# Add V1 endpoints

In springboot multiple request mappings can't be added to one method or class.

As the existing controller resource classes can't be mapped to `/` & `/v1` at the same time I have created V1 versions of each of the controller resource classes with corresponding methods that pass the request on to the original.

Once the UI has been updated to make requests to the `/v1` endpoints the original controller resource classes can be updated to have that request mapping and these V1 versions can be removed.