Async rekognition calls (#122)
Adds async versions of methods related to rekognition service. This is done by switching to `RekognitionAsyncClient`. Removes the blocking versions.

This branch was deployed in dev to verify it works.

Note: It would also make sense to use async S3 client here, but I wanted to keep this PR small.