Improve reporting accuracy by consolidating all calculations on the backend (#127)
ESUP-909

works with the equivalent branch in the UI repo

There was a number of calculations being performed in the UI repo, within the dashboard view itself, and some within the controller. This was identified as an issue so this PR moves the calculations (mostly total and percentage calculations) to the API to handle.